Node class re-structuring. Added delay node.
